Add deserialization backwards compatibility for n_audio (#60)
In `534d1b6`, the XcParams struct was updated to use a slice for storing audio indices.
However, this broke some use cases where the parameters struct was persisted by JSON
marshalling and then read back in, as the new code interpreted the audio index array to imply
a total of 8 audio indices.

This change adds a backwards compatibility layer to unmarshalling in order to fix that. It also
preserves the existing behavior of leaving fields on the XcParams that are not specified in any
place in the JSON as their previous values.