Make muxing timing adjustment based on dts
The muxing timing adjustments make it so that the output from muxing
starts at 0, and also doesn't end up being non-monotonic when the DTS
resets. This change makes it so that this works for cases where packets
are out of PTS-order, which previously would break because the PTS
would briefly double, and then raise errors because it was
non-monotonic.